Created new class "Product"
