+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Offer.java
@@ -22,9 +22,10 @@ public class Offer {
         return unavailableItems;
     }
 
+    //można zmodyfikować konstruktor, albo zrobić nowy konstruktor dostosowany do zmienionego obiektu
     @Override
     public int hashCode() {
-        final int prime = 31;
+        final int prime = 31;//
         int result = 1;
         result = prime * result + (availableItems == null ? 0 : availableItems.hashCode());
         return result;
