+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Product.java
@@ -1,0 +1,38 @@
+package pl.com.bottega.ecommerce.sales.domain.offer;
+
+import java.util.Date;
+
+public class Product {
+
+    private String productId;
+    private String productName;
+    private Date productSnapshotDate;
+
+    private String productType;
+
+    public String getProductName() {
+        return productName;
+    }
+
+    public Date getProductSnapshotDate() {
+        return productSnapshotDate;
+    }
+
+    public String getProductType() {
+        return productType;
+    }
+
+    public String getProductId() {
+        return productId;
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + (productName == null ? 0 : productName.hashCode());
+        result = prime * result + (productId == null ? 0 : productId.hashCode());
+        result = prime * result + (productType == null ? 0 : productType.hashCode());
+        return result;
+    }
+}
